// generated/api/gameFiltered.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * juniors bootcamp backend 🔥
 * Апи для выполнения индивидуальных заданий
 * OpenAPI spec version: 1.0
 */
import type { GameGenre } from "./gameGenre";
import type { GamePriceVariant } from "./gamePriceVariant";
import type { GameType } from "./gameType";

export interface GameFiltered {
  /** Slug игры */
  slug: string;
  /** Название игры */
  name: string;
  /** Дата релиза */
  releaseDate: number;
  /** Тип */
  type: GameType;
  /** Жанры игры */
  genres: GameGenre[];
  /** Изображение игры */
  image: string;
  /** Наименьший вариант цены */
  priceVariant: GamePriceVariant;
}